Reusable models
The user is able to save models after training (if there were more than
one R training session — that is, if the user clicked the Train link on
the Training side navigation bar more than once — the user can pick
which model to save).
It is possible to load previously saved models and use them on new data.
It is also possible to pick any model trained in a previous R session
and use it again on new data.