Selection based on classification results and class
The user is able to select data points based on the class it belongs to
or the classification result (TP, TN, FP, FN). In order to do this and
still and allow toggling the “confusion coloring mode” a context menu
is shown when the user clicks the rectangle on the legend listing the
classes.
Legend showing which color corresponds to which status (TP, TN, FP, FN)
was also added. It is through this legend that the user can select data
points based on the classification result.